# refactor: Improve textutil.PrefixString