This PR improves implementation of the `textutil.PrefixString` by removing unneeded `result` slice.