@AkihiroSuda 2024-11-05
> > What is improved?
> 
> The intermediate slice `result []string` is removed. This potentially speeds up. Also, -2 lines from the implementation.

How is that fast?