@afbjorklund 2024-11-05
If you wanted it "fast" it would probably use a Scanner, but for this particular feature I think bubble sort is just fine.