@alexandear 2024-11-05
> How is that fast?

I wrote benchmark:

```go
func BenchmarkPrefixString(b *testing.B) {
	text := "foo\nbar\nbaz\nqux\nquux\ncorge\ngrault\ngarply\nwaldo\nfred\nplugh\nxyzzy\nthud"
	prefix := "- "
	for i := 0; i < b.N; i++ {
		PrefixString(prefix, text)
	}
}
```

Run 10 times of it with old implementation:

```sh
$ go test -benchmem -bench=. -count=10 pkg/textutil/*.go > old.txt
```

and 10 times with new (this PR):
```sh
$ go test -benchmem -bench=. -count=10 pkg/textutil/*.go > new.txt
```

With the help of golang.org/x/perf/cmd/benchstat we can compare results:
```sh
$ benchstat old.txt new.txt
goos: darwin
goarch: arm64
cpu: Apple M1 Pro
               │   old.txt    │               new.txt               │
               │    sec/op    │   sec/op     vs base                │
PrefixString-8   1002.0n ± 3%   772.1n ± 9%  -22.94% (p=0.000 n=10)

               │  old.txt   │              new.txt               │
               │    B/op    │    B/op     vs base                │
PrefixString-8   896.0 ± 0%   400.0 ± 0%  -55.36% (p=0.000 n=10)

               │  old.txt   │              new.txt               │
               │ allocs/op  │ allocs/op   vs base                │
PrefixString-8   20.00 ± 0%   15.00 ± 0%  -25.00% (p=0.000 n=10)
```

`PrefixString` with the changes in this PR faster by 23%, allocates 55% less bytes/operation, and has 25% less allocations/operation.