@afbjorklund 2024-11-05
What I meant was that since we currently only indent two lines anyway, I don't think it matters?

Thanks for benchmarking, though!